Respect the clientId from the deltas and store the modified PK in deltas
If user1 exports the project at certain point of time, then creates a
feature in it, pushes the changes (not synced!), then modifies/deletes
the same feature and pushes that changes, the server was unable to find
what to modify/delete. Now it uses the clientId + localPk from the
delta, which is used as unique identifier.