Fix some flakiness in testing related to auditing
NOTE since we do not login via the API, the django_currentuser module is
unable to set the current user from the request object. Therefore the
user set from another test remains in the internal django_currentuser
cache and causes the test to fail. So we have to manually reset the
current user to None.