Skip downloading attachments in workers
Attachments are unimportant files such as photos, video, documents etc
that do not affect the work of the geospatial project.

- Add `is_attachment` flag to the files/package response to indicate
whether the file is considered an attachment.
- Do not download files within the worker that are flagged as `is_attachment`
- List files from the package and add the files from the original project
files in the attachment directories.
- When a file is requested from the package files endpoint and has an
attachment prefix, first check if the file is available in the package
and if not, try to download it from the original project files.