Fix replacing null bytes without utf8 conversion
If we decode to utf8, we might get a chunk splitting the utf8 char in two,
therefore creating invalid utf8 character.

If we work with bytes directly, we safely remove the null bytes without
worrying too much, as null bytes appear only for null characters in utf8.
Or at least according to this: https://stackoverflow.com/a/6907327/1226137