Refactor `delete_project_file_permanently` so the S3 storage is source of truth
The file operations depend on HTTP calls to the S3 storage and they might fail,
We need to choose source of truth between DB and S3.
For now the source of truth is what is on the S3 storage,
because we do most of our file operations directly by calling the S3 API.
1. S3 storage modification. If it fails, it will cancel the transaction
and do not update anything in the database.
For simplicity we now assume S3 storage is transactional, but we should be aware it might not hold be true all the times (e.g. request sent, processed, but response lost).
2. DB modification. If it fails, the states betewen DB and S3 mismatch,
but can be easyly synced from the S3 to DB with a manual script.