Prefer `ForeignKey(User, limit_choice_to=models.Q(type=Person))`
In short, it make community version of QFieldCLoud to prevent adding new collaborators or app.qfield.cloud to transfer organization to a new owner.

Turns out Django is not always aware that User and Person instances are almost the same, which causes various errors.
For example `Form.full_clean()` causes in `ModelForm` for `Organization` with `organization_owner` field.
It thinks there should be a `user_ptr` field, even though there is not.

Check https://opengisch.sentry.io/issues/3899409423/ .

Fix #514
Fix #515