Users getting error 500 when trying to register with username taken by an Organization
NOTE `allauth` depends on the `PRESERVE_USERNAME_CASING` to make `iexact` lookups.
When set to `False` as it is now, all `Person` usernames are stored as lowercase.
However, usernames for `Organization` or `Team` are not stored lowercase and then `allauth` does not check for them.

TODO check what are the consequences if we don't add all of these and change the `PRESERVE_USERNAME_CASING` to `True`.