Speed up delta_apply jobs by using an indexed column for searching
Filtering by `content__clientId` is causing super slow query. While in
pure SQL we can index on specific field in the JSON, like:

```CREATE INDEX ON core_delta(((content->>'clientId')::uuid))```

It is not so easy to run a query that uses the index in Django ORM.

Alternatively, we can create index directly on the JSON value:

```CREATE INDEX ON core_delta((content->'clientId'))```

But this will make the index 5x bigger.

In summary, it was chosen to add a new indexed column with the same data
as in the JSON, because it's storage cheaper and faster solution.