Create DEVELOPMENT.md
I tried to put some of the rules which have been followed to greater or lesser extent in the past of QFieldCloud development. I think pretty much everything has been discussed with the team in the past, but I think it would be nicer if we have it in written form.

Feel free to add suggestions, rewording or restructuring.

This document should be referred from `README.md` and most of the contents of the README should be moved here. Planning to do so in a follow-up PR.