Fix getting the list of available teams per organization
The problem comes from the fact that `teamname` is a runtime property of
the `Team` class and not an actual database  field.
In addition we were using `Team.object.values()` which returns an
`Iterable[dict]`, instead of `Iterable[Team]`. So even if it returned
the right type, it would have been terribly slow, due to N+1 situation.

This problem made the `/user/<username>/organization` return an error,
therefore prevented QFieldSync from selecting the available
organizations for the team.