Add Google and Microsoft as potentially supported IDPs:
This adds Google and Microsoft as potentially supported OpenID Connect
providers. It does not yet enable them, so they can't be used to log in yet.

Enabling them will require to add them to the SOCIALACCOUNT_PROVIDERS setting,
and define their configuration (client ID, etc..). Or add a configuration for
them as a SocialApp in the database, through the admin panel.

So this change just prepares QFC to possibly handle these two providers, once
we decide to add configuration for them in a particular deployment.