Customize admin login page to offer social logins:
This is a combination standard admin login page template from Jazzmin, plus
the snippet from django-allauth that adds the "Or use a third-party" with the
provider list at the bottom.

The (potential) SSO login options for the Jazzmin admin login page are
hidden behind a ?sso=1 query string parameter.

Doing this leaves the standard admin login page unchanged, and allows us for
a soft rollout of the feature. This is not a security measure at all, but a
"feature flag", intended to not accidentally confuse users / admins while the
feature is still in development.