Fix manual write to audit, usually about file changes, that is a string
In older version of audit, the value must be a string. This is no longer
the case and we have to push the value native type, usually a dict.

A manual patch of the database is needed, please run something like this:
```
update auditlog_logentry set changes = replace(replace(rtrim(ltrim(changes::text,'"'),'"'), '\"', '"'), '\\"', '')::jsonb where jsonb_typeof(changes) = 'string';
```