Unhide socialaccount models from admin panel:
Since the SocialAccount related models from django-allauth are now actually
used in QFC, they should not be hidden in the admin panel any more.

This allows us to easily inspect / debug linked social accounts for users.

It also lets us create SocialApp configurations through the web (in the
Django database) as an alternative to defining them in settings.py.