Add adapter that logs stack trace on 3rd party authentication errors:
The default django-allauth "Third-Party Login Failure" page is
decidedly unhelpful, because it is completely devoid of any error details.

And when errors happen in the callback view from the external provider, they
are usually quite tricky to debug already.

This adapter therefore makes sure the error details are logged, and especially
the stack trace is logged as well.

It also adds the stack trace to the template context. This would allow the
stack trace to be displayed in the socialaccount/authentication_error.html
template. We don't do this yet, because in production this could pose a
security risk (displaying stack traces to users, especially authentication
related ones). But in scenarios where this is deemed acceptable (some on-prem
deployments for example) the information is there, and can be displayed with
a simple template override.