Set SOCIALACCOUNT_QUERY_EMAIL = True:
This makes sure that the SocialAccount providers always get queried for the
user's email address, among other needed scopes.

The email address is required for the way we want to link social accounts to
possibly already existing user accounts in QFC, and setting this globally
avoids having to explicitly set this in the scope when configuring the
provider (which is an easy mistake to make, and leads to strange results).