Set SOCIALACCOUNT_EMAIL_AUTHENTICATION = True:
This will (globally) allow any social account providers to be used to
authenticate with an existing user based on the email address that the
social account provider reports.

This has the following security implication:
If we ever configure and allow a malicious or negligent social account
provider to be used, that provider (IDP), or one of its users, could take over
an existing QFC account.

There are two scenarios to consider:

1) The provider itself is malicious.
The provider could then spoof any email address they want in the userinfo
they return to us, we trust it, and allow them to authenticate with an existing
user account that matches that email address.

2) The provider is negligent, and doesn't validate email addresses.
A user of that provider could sign up with the IDP using a fake email address
that doesn't get validated by the provider. The user would then be able to
authenticate as an existing QFC user with that email address.

Having said that, the big players  in terms of OpenID Connect IDPs (Google,
Microsoft, GitHub, ...) all do validate their user's email addresses (or even
control them). Therefore I think the risk of setting this globally (instead
of per-provider) is acceptable, compared to the risk of forgetting to set
EMAIL_AUTHENTICATION for every configured provider.