Set SOCIALACCOUNT_EMAIL_AUTHENTICATION_AUTO_CONNECT = True:
This makes sure that users are not only able to authenticate using a social
account (via a matching email address), but that an existing QFC user account
then automatically gets linked with that SocialAccount.

This has the following side effect:
If the existing QFC user account has an email address that has NOT been
verified, the accounts password will be scrambled (made unusable).

The reason for this is a security concern:
This prevents "account squatting" - where a malicious actor would first
register with an email address of the victim, QFC doesn't validate that email,
and the malicious actor gets to set his own password.
When the victim then at some later point logs in using a social account, using
that email address, and his account gets auto-linked, then the attacker would
have a backdoor to the legitimate owner of that email address's QFC account,
by means of the password.

QFC currently does do email address validation. When checking in production,
we found that only 0.1% of users didn't have a validated email address, and
they didn't seem to be primary emails. In addition, password reset is possible
once the user logs into QFC using the social account. So their password
getting scrambled wouldn't be the end of the world, just a minor annoyance.