Set SOCIALACCOUNT_LOGIN_ON_GET = True:
This skips the secondary form that requires you to confirm that you *really*
want to log in using <provider>, by not requiring POST requests to trigger
social logins.

The django-allauth docs mention a security concern about this, but after
looking into it, I think it's pretty much a non-issue:

The scenario they describe in the django-allauth 0.47.0 release notes assumes
that there is what's called an "open redirect" vulnerability on the IDP side.
That is a major, and long known security vulnerability that no big player like
Google, Facebook or Microsoft should have. I therefore think we can safely
enable this.