Add QGIS authentication middleware:
This middleware allows to authenticate to QFC using QGIS auth configs with an
OpenID Connect Authorization Code flow.

The reason this middleware is needed is as follows:
QGIS already has support for performing an OAuth2 Authorization Code flow
using its authentication configs.

However, the way that QGIS passes the result of that flow (an access_token) to
QFC is very limited, and cannot really be controlled. QGIS just runs the flow
for the user, and then attaches the access_token it gets to any subsequent
requests in the `Authorization` header.

However, django-allauth, for a regular OIDC flow,  expects the final request
of the flow to hit e.g. /accounts/google/login/callback, then will log the
user in and create a session.

Because QGIS doesn't hit *any* particular URL first, like a callback URL, but
instead just attaches the token to every request, we need a middleware that
looks for those tokens in any arbitrary requests, and authenticates them if
necessary.

Once the authentication is performed (preferably using an OIDC ID token, or
as a a fallback an OAuth2 access token), a session is created, and on
subsequent requests the user is already authenticaten (since QGIS sends back
the session cookies it gets).