Add API endpoint to list enabled auth providers:
This introduces a new API endpoint api/v1/auth/providers/ that allows to
enumerate all the supported and enabled authentication providers in QFC.

This can be used by QFC clients (apps like QField or the QFieldSync plugin)
to present the user with the appropriate options to log in.

The information returned from this endpoint contains the necessary details
for a QFC client to build an appropriate QGIS auth config to log in using
the selected provider.