auth/providers endpoint: Include style information for providers:
This includes a new `styles` property for each provider returned from the
endpoint. This is needed for QFC clients (like QField or the QFieldSync plugin)
to render the SSO login options with a logo and appropriate brand colors.

The returned styles include two themes, "dark" and "light", which contain
appropriate colors and logos for use in light or dark modes.

These styles are controlled via a new setting QFIELDCLOUD_SSO_PROVIDER_STYLES,
which is currently prefilled for the providers we know we'll need in the near
future (Google, GitHub and Keycloak).