Inject provider styles into d.allauth templates:
This adds the provider style information onto django-allauth's `Provider`
objects when they get listed. This allows for them to be easily accessed
in HTML templates, where those styles are needed to customize the rendering
of the provider list.

We also override django-allauth's `provider_list.html` template in order to
pass the `styles` dictionaries along to the `provider.html` snippet, which
renders an individual provider button.