feat(secrets)!: add support for per user project secrets
So far QFieldCloud supported secrets that are stored on per-project basis.
So if there are multiple collaborators in one project, they all reuse the same secrets.

This commit allows in addition of the per project settings a user to be
able to set their own secrets.

A sample scenario when this is used:
An organization admin has 2 users in the same project with PostGIS layer
, but they need to see different subset of a particular offlined layer.
So far when both users will share the same credentials provided by the
organization admin as a PGSERVICE file.

With the changes introduced in this PR, the organization admin can
provide separate PGSERVICE file for each user, therefore automatic
filtering can happen already in the PostGIS database based on the DB user.

BREAKING CHANGE: introduces a migration that adds new fields to the
`Secret` model, creates unique indices and on reverse deletes data.