refactor(localized): rename all `localized_` prefixed variables to `shared_` regarding shared projects
Also rename the project name from `localized_datasets` to `shared_datasets`.

Improves the readability and explainability of the feature.

Note: there is a "breaking change" in the sense that `Project.localized_datasets_project_id` was renamed to `Project.shared_datasets_project_id` in the serializer, but the clients do not support this feature officially yet anyways, so it is acceptable.