fix: calculate the `DynamicStorageFieldFile.storage` on demand in a cached property
The new approach removes the pre-calculated in the `__init__` value of
`DynamicStorageFieldFile.storage` attribute, and adds a dynamic property
with the same name. For further performance boost, the property is cached.
This way we still calculate the value only when it is needed.

Tested with legacy, non-legacy and combination of legacy and non-legacy
storage locally.

reverts the changes added as of `a026983c920152a7e9e8f3313cf07daeeb75d436`,
but still calculate the `storage` on demand. The quoted commit caused:
```
  File \"/usr/src/app/qfieldcloud/core/fields.py\", line 99, in pre_save
    file = super().pre_save(model_instance, add)
  File \"/usr/local/lib/python3.10/site-packages/django/db/models/fields/files.py\", line 317, in pre_save
    file.save(file.name, file.file, save=False)
  File \"/usr/local/lib/python3.10/site-packages/django/db/models/fields/files.py\", line 93, in save
    self.name = self.storage.save(name, content, max_length=self.field.max_length)
AttributeError: '__proxy__' object has no attribute 'save'"
```