refactor: move the `File.file_storage` to `FileVersion.file_version` model
First we restore the initialization of `DynamicStorageFieldFile.storage` in `DynamicStorageFieldFile.__ini__` and we forget about all the workarounds with (cached) dynamic properties.

Second, since the data is already within the `FileVersion` record, we save the N+1 SQL query that was getting the `file_storage` via additional `File.file_storage` request for each version on initialization.