Add dedicated model to store faulty deltafiles:
This introduces a new model FaultyDeltaFile that is used to preserve invalid
deltafiles that we failed to parse or process.

When an error happens during parsing a deltafile's JSON, or processing it,
the exception is caught and logged. In addition, the contents of the
offending deltafile are stored in a FaultyDeltaFile instance, together with
the traceback and some metadata that should help debug it.