fix(nginx): CSRF check failure when debug mode is enabled
Steps to reproduce:

1) Make a potential runtime error in the codebase, e.g. add to the Project.save something that raises an unhandled error, such as 1/0
2) Go to admin and try to save an existing/new instance of the Project model.
3) We would expect an error 500 to happen and override it with the 500.html template
4) We experience CSRF verification failure instead.

What happens is that nginx detects the error 500, then it continues with error_page 500 501 505 = @handle_500_error;  which eventually proxies back to Django if in debug mode. But there were no proxy_set_* configurations, so basically Django got a second request, which is HTTP 1.0 instead of HTTP 1.1 and missing all the X-Forwarded-*  headers. See the tcpflow attached for more info.

This second request is actually wrong, it should not happen and it is very dangerous, as error 500 might cause damage and sending a request twice might cause even bigger damage! Luckily this is only for debug mode, but we need another solution to get the original response from Django instead of the debug page.