Use variable storage key for thumbnails:
Using a variable storage key that changes every time for thumbnails is an
easy way to avoid thumbnail files exceeding the maximum allowed versions
on object storages with versioning enabled.

Before, when a fixed storage key like 'projects/{proj}/meta/thumbnail.png'
was used, this created a new version on the object storage level for every
new thumbnail. This has lead to situations where that thumbnail file then
exceeded the maximum amount of 1000 versions, and caused issues.

By using a variable key we save every new thumbnails as a new object, with
always just a single version. When project.thumbnail is updated, and the
reference to the old thumbnail file is removed, the old file will be marked
as deleted. These deleted files can in the future then be cleaned up with
a script or a maintenance job.