feat(worker): add a new `QfcBaseWorkflow.get_workflow` method
Instead of having `QfcBaseWorkflow.handle` to take care of building the
workflow and making sure the `feedback.json` is generated, we make the
`handle` still prepare the `feedback.json`, but now in the base class
and the new method `get_workflow` must be implemented in the children.

This also allow for easier extension of already existing jobs.