feat: kill the geodb once and forever
Geodb used to be a core functionality on pre-production days of QFC,
where users can create their own postgis database dynamically and this
could be sold as additional product along QFieldCloud.

This never took off due to complexities of managing a database for
someone, but the testing framework was heavily dependent on the GeoDB
already.

So long Geodb and thanks for the fish!

RELEASE NOTES:

- When starting the stack, make sure you restart it with `--remove-orphans` flag to delete the orphaned `geodb` service container. This is valid for both local development, staging and testing.
- Remove all `GEODB_*` envvars from `.env` file.
- Applying the change will permanently delete the core_geodb table. If you still somehow use Geodb, do not use this version.