refactor: no need for the `WEB_HTTP(S)_PORT` in the `app` container
Not sure what the original author (me from the past) thought with the
direct check of the port, since `request.is_secure()` checks exactly
this and in the current docker compose stack we always use `nginx`.

If someone decides to change the front from nginx to something else,
then we will encounter a problem in the current or the latter case, so
there was no additional safety.