Use only one login page from `allauth`
Since login is a quite a critical part of every application, it would be great if it is always done via a single entrypoint. However, currently the login is handled either via allauth login, or via Django admin login, or Django's built in auth when requested via the browsable DRF API endpoint.

This change unifies all authentication flows to use the `allauth` login page, ensuring a consistent user experience and future compatibility with SSO.