Tests: Fix flaky files API tests:
Some of these tests occasionally failed in a flaky way. The reason was that
methods in QfcFilesTestCaseMixin created AuthTokens using this pattern:

AuthToken.objects.get_or_create(user=user)

This created a token for the user, with the client type UNKNOWN, because a
user agent is never set in tests. This worked, unless during a test a worker
job was triggered as a side effect.

Then the worker wrapper would issue a token for the same user, but with
client_type WORKER. This would then cause get_or_create(user=user) to fail,
because (as per docs) it assumes that its keyword arguments act as a unique
constraint for the object to be fetched.

Further discriminating the get_or_create() call by the client_type addresses
this issue.