feat: add support for virtual layers without embedded layers
Since the virtual and memory layers do not have filenames, we should not
treat them as regular postgres or WFS layers when checking for `Project.
has_online_vector_data`.

On the other hand, the virtual layers may embed a postgresql layer and
we should disable this workaround when a `Plan.is_external_db_supported`
is `True`.

We take a very radical approach. Instead of looking at the current
embedded layer provider, we just say - if there are any embedded layers
(which may be file based and part of the project already), we do assume
the layer has online vector data and therefore external db.