Temporarily bump django-auditlog to version from git that contains SSO fix:
This uses django-auditlog from git, pinned to a commit that contains a fix we
need to prevent a bug during SSO login.

See jazzband/django-auditlog#784

This change also adds `git` as a dependency to the app Dockerfile, since it's
now needed during `pip install`.

Once the next version of django-auditlog is released (e.g. 3.3.1), this should
be reverted back to a regular version, and `git` should be removed from the
Dockerfile again.