Define AUDITLOG_LOGENTRY_MODEL in settings:
This sets the new auditlog setting AUDITLOG_LOGENTRY_MODEL to its default
value of "auditlog.LogEntry" in settings.

This is necessary because on current master of `django-auditlog` this new
setting was added a couple weeks ago. And depending on the exact loading
order in INSTALLED_APPS, Django might trigger module autodiscovery from
django.contrib.admin *before* auditlog has had a chance to initialize this
setting with its default value.

This could potentially be avoided by changing the position of `auditlog` in
INSTALLED_APPS, but that seems like a much riskier change. Explicitly adding
this setting with its default value to our configuration makes it independent
of the exact loading order, and seems safer to me.