feat: add object storage cleaner script and testing
Currently QFieldCloud keeps all project files using the non-legacy storage even if they are deleted by the user.
This is due to enabled versioning on Exoscale and the lack of Exoscale mechanism to delete files after certain amount of time after they are deleted.
This causes excessive file storage and therefore costs.

- Add `storage_cleaner.py` for scanning and deleting logically deleted objects in S3 compatable storages
- Add `test.py` for testing the script