Introduce `flake8-if-expr` to forbid inline `if`s
We always had trouble to explain what is the tolerance of using inline `if`s.

Now with this flake8 rule, there is no longer arguemntation if something
subjectively passes.

I am quite happy with the readability improvement. As someone that
switches languages all the time, this is a huge leap forward.